requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
openai>=1.40.0
python-dotenv>=1.0.1
langgraph>=0.2.0
pydantic>=2.6
psycopg[binary]>=3.1
chromadb>=0.5.0
